Force query dispatcher to always use gpsegconfig_dump file
Introduce GUC gp_force_gpsegconfig_dump_usage to force the query
dispatcher to always use the gpsegconfig_dump file. This file is
normally used in 2PC retry scenarios in a normally operating GPDB to
prevent deadlocking FTS and thus bypassing it.

For Hot Standby Cluster, this mechanism can instead be retrofitted to
allow the Hot Standby master segment to query dispatch without
modifying the gp_segment_configuration (which is not possible on a hot
standby since it's read-only).