Update distributed commit WAL replay for hot standby
The distributed commit WAL is essentially the same as regular commit
WAL so cache invalidation messages exist. These need to be parsed and
processed so that caches are properly invalidated for hot standby
connections to the GPDB standby master (or any replica of the
master). To do this, we copy over similar logic from the more
maintained xact_redo_commit function.