Do not wait for synchronous replication if QD WAL sender is not for HA
In GPDB, the master segment does not follow the usual rules for
synchronous replication. It just assumes that all WAL streaming
connections are synchronous no matter if the synchronous_standby_names
GUC is set or not (to allow the master to skip blocking if the standby
master crashes). To allow asynchronous connections to the master
segment, we should check for the is_for_gp_walreceiver flag that we
label our HA mirror connections with. If it's not set to true, we
should not consider the WAL stream to be synchronous.