Update XLogCtl->recoveryLastXTime for distributed commit record replay
Querying pg_last_xact_replay_timestamp() would give a NULL result on
the master segment. This is because we do not update
XLogCtl->recoveryLastXTime when a distributed commit record is
replayed. Just need to add it to the conditional check in
getRecordTimestamp().

One thing to note is that distributed forget records are not tracked
in this logic because they do not have transaction timestamps. We
could probably use the distributed timestamp by cracking the gid
recorded in the record but not sure if that would be valid yet.